Added s2_wofs
