+++ workfinder/search/ml_water.py
@@ -21,20 +21,6 @@ class BaseMLWater(BaseWofs):
     def __init__(self, s3: S3Api, redis: BaseQueue):
         super().__init__(s3, redis)
 
-    def submit_tasks(self, to_do_list: pd.DataFrame):
-        region = get_config("APP", "REGION")
-        target_bucket = get_config("S3", "BUCKET")
-        target_queue = get_config("ML_WATER", "REDIS_PROCESSED_CHANNEL")
-        wofs_summary = get_config("ML_WATER", "WOFS_SUMMARY_PATH")
-        imagery_path = get_config("S3", "IMAGERY_PATH")
-        for r in to_do_list.tolist():
-            payload = {
-                "img_yml_path": r['url'],
-                "lab_yml_path": wofs_summary,
-                "s3_bucket": target_bucket,
-                "s3_dir": f"{imagery_path}/{region.lower()}/{self.get_target_name()}/"
-            }
-            self._redis.publish(target_queue, json.dumps(payload))
 
 
 class Landsat8MLWater(BaseMLWater):
